requirements.txt
asgiref==3.8.1
attrs==24.2.0
Django==5.1.3
django-extensions==3.2.3
django-filter==24.3
django-restframework==0.0.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
factory_boy==3.3.1
Faker==33.1.0
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
psycopg2==2.9.10
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.21.0
six==1.17.0
sqlparse==0.5.2
typing_extensions==4.12.2
uritemplate==4.1.1
